requirements.txt
slack_bolt
requests
psycopg2-binary
